+++ packages/smooth_app/lib/pages/preferences/user_preferences_haptic_feedback.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:provider/provider.dart';
+import 'package:smooth_app/data_models/preferences/user_preferences.dart';
+import 'package:smooth_app/pages/preferences/user_preferences_widgets.dart';
+
+class UserPreferencesHapticFeedback extends StatelessWidget {
+  const UserPreferencesHapticFeedback();
+
+  @override
+  Widget build(BuildContext context) {
+    final AppLocalizations appLocalizations = AppLocalizations.of(context);
+    final UserPreferences userPreferences = context.watch<UserPreferences>();
+    return UserPreferencesSwitchItem(
+      title: appLocalizations.app_haptic_feedback_title,
+      subtitle: appLocalizations.app_haptic_feedback_subtitle,
+      value: userPreferences.hapticFeedbackEnabled,
+      onChanged: (final bool value) async =>
+          userPreferences.setHapticFeedbackEnabled(value),
+    );
+  }
+}
